+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test-watch": "node_modules/.bin/rerun-script test",
     "test": "clear; npm run lint; npm run unit",
-    "unit": "NODE_PATH=./lib lab -c -v -m 3000 -f test -a code",
+    "unit": "NODE_PATH=./lib lab -c -v -m 3000 -a code ./test/unit",
     "start": "NODE_PATH=./lib node index.js",
     "format": "standard --format --verbose",
     "format-snazzy": "standard --format --verbose | snazzy",
